\name{convertBCD-package}
\alias{convertBCD-package}
\alias{convertBCD}
\docType{package}
\title{
\packageTitle{convertBCD}
}
\description{
\packageDescription{convertBCD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{convertBCD}
\packageIndices{convertBCD}

}
\author{
\packageAuthor{convertBCD}

Maintainer: \packageMaintainer{convertBCD}
}

\keyword{ package }

